ref(conversations): Remove redesign feature flag and old components
The gen-ai-conversations-redesign rollout is complete, so make the
redesigned components the only path and drop the flag checks throughout the
conversations product area and the trace AI drawer.

Delete the superseded flag-off components. Helpers that the redesigned
components still rely on are moved into their New counterparts so no module
survives purely to re-export them, keeping the follow-up rename mechanical.

Refs TET-2742
