+++ static/app/views/explore/conversations/components/conversationLayout.tsx
@@ -4,7 +4,6 @@ import {useRef} from 'react';
 import {Container, Flex, Stack} from '@sentry/scraps/layout';
 import {SplitPanel} from '@sentry/scraps/splitPanel';
 
-import {Placeholder} from 'sentry/components/placeholder';
 import {useDimensions} from 'sentry/utils/useDimensions';
 import {useLocalStorageState} from 'sentry/utils/useLocalStorageState';
 import {useOrganization} from 'sentry/utils/useOrganization';
@@ -242,80 +241,5 @@ export function ConversationDetailPanel({
         initiallyCollapseAiIO,
       })}
     </Stack>
-  );
-}
-
-export function ConversationViewSkeleton() {
-  return (
-    <ConversationSplitLayout
-      left={
-        <ConversationLeftPanel>
-          <Container borderBottom="primary" padding="md lg">
-            <Flex gap="lg">
-              <Placeholder height="14px" width="40px" />
-              <Placeholder height="14px" width="40px" />
-            </Flex>
-          </Container>
-          <Stack flex="1" gap="md" padding="lg" background="secondary">
-            <Stack gap="sm" padding="sm md">
-              <Placeholder height="12px" width="120px" />
-              <Placeholder height="12px" width="80%" />
-            </Stack>
-            <Container background="primary" radius="md" border="primary" padding="sm md">
-              <Stack gap="sm">
-                <Flex align="center" gap="sm">
-                  <Placeholder height="12px" width="100px" />
-                  <Placeholder height="12px" width="40px" />
-                </Flex>
-                <Container background="tertiary" radius="sm" padding="xs sm">
-                  <Placeholder height="12px" width="150px" />
-                </Container>
-                <Placeholder height="12px" width="90%" />
-                <Placeholder height="12px" width="70%" />
-                <Placeholder height="12px" width="60%" />
-              </Stack>
-            </Container>
-            <Stack gap="sm" padding="sm md">
-              <Placeholder height="12px" width="120px" />
-              <Placeholder height="12px" width="60%" />
-            </Stack>
-            <Container background="primary" radius="md" border="primary" padding="sm md">
-              <Stack gap="sm">
-                <Flex align="center" gap="sm">
-                  <Placeholder height="12px" width="80px" />
-                  <Placeholder height="12px" width="35px" />
-                </Flex>
-                <Placeholder height="12px" width="85%" />
-                <Placeholder height="12px" width="50%" />
-              </Stack>
-            </Container>
-          </Stack>
-        </ConversationLeftPanel>
-      }
-      right={
-        <Stack gap="lg" padding="lg">
-          <Stack gap="sm">
-            <Placeholder height="14px" width="180px" />
-            <Placeholder height="16px" width="60px" />
-          </Stack>
-          <Stack gap="sm">
-            <Placeholder height="12px" width="80px" />
-            <Placeholder height="12px" width="200px" />
-          </Stack>
-          <Stack gap="sm">
-            <Placeholder height="12px" width="60px" />
-            <Placeholder height="12px" width="160px" />
-          </Stack>
-          <Stack gap="sm">
-            <Placeholder height="14px" width="80px" />
-            <Placeholder height="80px" width="100%" />
-          </Stack>
-          <Stack gap="sm">
-            <Placeholder height="14px" width="80px" />
-            <Placeholder height="120px" width="100%" />
-          </Stack>
-        </Stack>
-      }
-    />
   );
 }
