--- static/app/views/explore/conversations/components/messagesPanel.spec.tsx
@@ -1,648 +1,0 @@
-import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
-
-import {EMPTY_TEXT_CONTENT} from 'sentry/views/insights/pages/agents/utils/aiMessageNormalizer';
-import {SpanFields} from 'sentry/views/insights/types';
-
-import {MessagesPanel} from './messagesPanel';
-
-function createMockNode(overrides: {
-  id: string;
-  attributes?: Record<string, string | number>;
-  endTimestamp?: number;
-  startTimestamp?: number;
-}) {
-  const {id, attributes = {}, startTimestamp = 1000, endTimestamp} = overrides;
-  const end = endTimestamp ?? startTimestamp + 100;
-  return {
-    id,
-    type: 'span' as const,
-    op: 'gen_ai.generate',
-    startTimestamp,
-    endTimestamp: end,
-    value: {
-      start_timestamp: startTimestamp,
-      end_timestamp: end,
-    },
-    attributes: {
-      // Must be 'ai_client' for getIsAiGenerationSpan to return true
-      [SpanFields.GEN_AI_OPERATION_TYPE]: 'ai_client',
-      ...attributes,
-    },
-    errors: new Set(),
-  };
-}
-
-function createMockToolNode(overrides: {
-  id: string;
-  toolName: string;
-  endTimestamp?: number;
-  startTimestamp?: number;
-}) {
-  const {id, toolName, startTimestamp = 1000, endTimestamp} = overrides;
-  const end = endTimestamp ?? startTimestamp + 100;
-  return {
-    id,
-    type: 'span' as const,
-    op: 'gen_ai.execute_tool',
-    startTimestamp,
-    endTimestamp: end,
-    value: {
-      start_timestamp: startTimestamp,
-      end_timestamp: end,
-    },
-    attributes: {
-      [SpanFields.GEN_AI_OPERATION_TYPE]: 'tool',
-      [SpanFields.GEN_AI_TOOL_NAME]: toolName,
-    },
-    errors: new Set(),
-  };
-}
-
-describe('MessagesPanel', () => {
-  const mockOnSelectNode = jest.fn();
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('renders empty message when no nodes provided', () => {
-    render(
-      <MessagesPanel nodes={[]} selectedNodeId={null} onSelectNode={mockOnSelectNode} />
-    );
-
-    expect(screen.getByText('No messages found')).toBeInTheDocument();
-  });
-
-  it('renders user and assistant messages from gen_ai.request.messages', () => {
-    const requestMessages = JSON.stringify([
-      {role: 'user', content: 'Hello from request messages'},
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Assistant response text',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Hello from request messages')).toBeInTheDocument();
-    expect(screen.getByText('Assistant response text')).toBeInTheDocument();
-  });
-
-  it('prefers gen_ai.output.messages over gen_ai.response.text for assistant', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-    const outputMessages = JSON.stringify([
-      {role: 'assistant', content: 'Hello from OUTPUT messages'},
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_OUTPUT_MESSAGES]: outputMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Fallback response text',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Hello from OUTPUT messages')).toBeInTheDocument();
-    expect(screen.queryByText('Fallback response text')).not.toBeInTheDocument();
-  });
-
-  it('renders a placeholder when output text content is missing', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-    const outputMessages = JSON.stringify([
-      {role: 'assistant', content: [{type: 'text', chars: 56}]},
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_OUTPUT_MESSAGES]: outputMessages,
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText(EMPTY_TEXT_CONTENT)).toBeInTheDocument();
-  });
-
-  it('prefers gen_ai.input.messages over gen_ai.request.messages', () => {
-    const inputMessages = JSON.stringify([
-      {role: 'user', content: 'Hello from INPUT messages'},
-    ]);
-    const requestMessages = JSON.stringify([
-      {role: 'user', content: 'Hello from REQUEST messages'},
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_INPUT_MESSAGES]: inputMessages,
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Hello from INPUT messages')).toBeInTheDocument();
-    expect(screen.queryByText('Hello from REQUEST messages')).not.toBeInTheDocument();
-  });
-
-  it('falls back to gen_ai.request.messages when gen_ai.input.messages is empty string', () => {
-    const requestMessages = JSON.stringify([
-      {role: 'user', content: 'Hello from REQUEST messages fallback'},
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_INPUT_MESSAGES]: '',
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Hello from REQUEST messages fallback')).toBeInTheDocument();
-  });
-
-  it('handles messages with parts format', () => {
-    const inputMessages = JSON.stringify([
-      {
-        role: 'user',
-        parts: [{type: 'text', content: 'Message from parts format'}],
-      },
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_INPUT_MESSAGES]: inputMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Message from parts format')).toBeInTheDocument();
-  });
-
-  it('handles messages with array content format', () => {
-    const inputMessages = JSON.stringify([
-      {
-        role: 'user',
-        content: [{text: 'Message from array content'}],
-      },
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_INPUT_MESSAGES]: inputMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Message from array content')).toBeInTheDocument();
-  });
-
-  it('calls onSelectNode when assistant message is clicked', async () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Click me assistant',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    // Find the assistant text and click its parent container
-    const assistantText = screen.getByText('Click me assistant');
-    // The clickable area is the MessageBubble which contains the text
-    // We can click on the text itself - the event will bubble up
-    await userEvent.click(assistantText);
-
-    expect(mockOnSelectNode).toHaveBeenCalledWith(node);
-  });
-
-  it('displays user email when available', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-        [SpanFields.USER_EMAIL]: 'test@example.com',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('test@example.com')).toBeInTheDocument();
-  });
-
-  it('deduplicates a last user message replayed across a cumulative tool loop', () => {
-    const node1 = createMockNode({
-      id: 'span-1',
-      startTimestamp: 1000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: JSON.stringify([
-          {role: 'user', content: 'Duplicate message'},
-        ]),
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response 1',
-      },
-    });
-
-    // Cumulative history replaying the same last message: shown once.
-    const node2 = createMockNode({
-      id: 'span-2',
-      startTimestamp: 2000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: JSON.stringify([
-          {role: 'user', content: 'Duplicate message'},
-          {role: 'assistant', content: 'Response 1'},
-        ]),
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response 2',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node1, node2] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    // Should only show one instance of the duplicate user message
-    expect(screen.getAllByText('Duplicate message')).toHaveLength(1);
-    // But both assistant responses should be shown
-    expect(screen.getByText('Response 1')).toBeInTheDocument();
-    expect(screen.getByText('Response 2')).toBeInTheDocument();
-  });
-
-  it('displays agent name in assistant header when available', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-        [SpanFields.GEN_AI_AGENT_NAME]: 'my-cool-agent',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('my-cool-agent')).toBeInTheDocument();
-    expect(screen.queryByText('Assistant')).not.toBeInTheDocument();
-  });
-
-  it('displays model name in assistant header when no agent name', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-        [SpanFields.GEN_AI_RESPONSE_MODEL]: 'gpt-4o',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('gpt-4o')).toBeInTheDocument();
-    expect(screen.queryByText('Assistant')).not.toBeInTheDocument();
-  });
-
-  it('falls back to Assistant when no agent name or model name', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Response',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Assistant')).toBeInTheDocument();
-  });
-
-  it('displays tool calls on assistant messages', () => {
-    const requestMessages = JSON.stringify([
-      {role: 'user', content: 'Check the weather'},
-    ]);
-
-    // Generation span that triggers tool calls
-    const generationNode1 = createMockNode({
-      id: 'span-1',
-      startTimestamp: 1000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Let me check the weather for you',
-      },
-    });
-
-    // Tool execution spans
-    const toolNode1 = createMockToolNode({
-      id: 'tool-1',
-      toolName: 'weather',
-      startTimestamp: 1500,
-    });
-
-    // Generation span with results
-    const generationNode2 = createMockNode({
-      id: 'span-2',
-      startTimestamp: 2000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'The weather is sunny',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[generationNode1, toolNode1, generationNode2] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('The weather is sunny')).toBeInTheDocument();
-
-    expect(screen.getByText('weather')).toBeInTheDocument();
-  });
-
-  it('renders tool calls with called tool prefix', () => {
-    const requestMessages = JSON.stringify([
-      {role: 'user', content: 'Check the weather'},
-    ]);
-
-    const generationNode1 = createMockNode({
-      id: 'span-1',
-      startTimestamp: 1000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Let me check',
-      },
-    });
-
-    const toolNode1 = createMockToolNode({
-      id: 'tool-1',
-      toolName: 'weather_api',
-      startTimestamp: 1500,
-    });
-
-    const generationNode2 = createMockNode({
-      id: 'span-2',
-      startTimestamp: 2000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'The weather is sunny',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[generationNode1, toolNode1, generationNode2] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Called tool')).toBeInTheDocument();
-    expect(screen.getByText('weather_api')).toBeInTheDocument();
-  });
-
-  it('keeps reasoning text in the DOM inside a collapsed details element', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-    const outputMessages = JSON.stringify([
-      {
-        role: 'assistant',
-        parts: [
-          {type: 'reasoning', content: 'My secret thinking text'},
-          {type: 'text', text: 'The final answer'},
-        ],
-      },
-    ]);
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_OUTPUT_MESSAGES]: outputMessages,
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    // Reasoning text is rendered in the DOM even though the <details> is
-    // collapsed, so the browser's find-in-page (Ctrl-F) can locate it and
-    // natively open the section.
-    const reasoning = screen.getByText('My secret thinking text');
-    expect(reasoning).toBeInTheDocument();
-    const details = reasoning.closest('details');
-    expect(details).not.toBeNull();
-    expect(details).not.toHaveAttribute('open');
-  });
-
-  it('handles output messages as a JSON object with content key', () => {
-    const requestMessages = JSON.stringify([{role: 'user', content: 'User message'}]);
-    const outputObj = JSON.stringify({content: 'Response from object format'});
-
-    const node = createMockNode({
-      id: 'span-1',
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_OUTPUT_MESSAGES]: outputObj,
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={[node] as any}
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    expect(screen.getByText('Response from object format')).toBeInTheDocument();
-  });
-
-  it('carries forward tool calls from spans without text to the next message with text', () => {
-    const requestMessages = JSON.stringify([
-      {role: 'user', content: 'Compare weather in Spain and Germany'},
-    ]);
-
-    // First generation span with text
-    const generationNode1 = createMockNode({
-      id: 'span-1',
-      startTimestamp: 1000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Let me check the weather for Spain',
-      },
-    });
-
-    // Tool execution spans (weather lookups)
-    const toolNode1 = createMockToolNode({
-      id: 'tool-1',
-      toolName: 'weather',
-      startTimestamp: 1500,
-    });
-    const toolNode2 = createMockToolNode({
-      id: 'tool-2',
-      toolName: 'weather',
-      startTimestamp: 1600,
-    });
-
-    // Generation span WITHOUT text (only made a tool call, no response text)
-    // This simulates when the LLM decides to call another tool without producing text
-    const generationNode2 = createMockNode({
-      id: 'span-2',
-      startTimestamp: 2000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        // No GEN_AI_RESPONSE_TEXT - simulates a span that only made tool calls
-      },
-    });
-
-    // Calculator tool execution
-    const toolNode3 = createMockToolNode({
-      id: 'tool-3',
-      toolName: 'calculator',
-      startTimestamp: 2500,
-    });
-
-    // Final generation span with text (shows the comparison results)
-    const generationNode3 = createMockNode({
-      id: 'span-3',
-      startTimestamp: 3000,
-      attributes: {
-        [SpanFields.GEN_AI_REQUEST_MESSAGES]: requestMessages,
-        [SpanFields.GEN_AI_RESPONSE_TEXT]: 'Here is the comparison',
-      },
-    });
-
-    render(
-      <MessagesPanel
-        nodes={
-          [
-            generationNode1,
-            toolNode1,
-            toolNode2,
-            generationNode2,
-            toolNode3,
-            generationNode3,
-          ] as any
-        }
-        selectedNodeId={null}
-        onSelectNode={mockOnSelectNode}
-      />
-    );
-
-    // The final message should show all tool calls (weather x2 from the skipped span + calculator)
-    expect(screen.getByText('Here is the comparison')).toBeInTheDocument();
-
-    // Should have 2 weather tags and 1 calculator tag
-    expect(screen.getAllByText('weather')).toHaveLength(2);
-    expect(screen.getByText('calculator')).toBeInTheDocument();
-  });
-});
