--- static/app/views/explore/conversations/components/messagesPanel.tsx
@@ -1,285 +1,0 @@
-import {useCallback, useMemo, useState} from 'react';
-import {css} from '@emotion/react';
-import styled from '@emotion/styled';
-
-import {Container, Flex, Stack} from '@sentry/scraps/layout';
-import {Text} from '@sentry/scraps/text';
-
-import {ClippedBox} from 'sentry/components/clippedBox';
-import {EmptyMessage} from 'sentry/components/emptyMessage';
-import {IconChevron} from 'sentry/icons';
-import {t} from 'sentry/locale';
-import {trackAnalytics} from 'sentry/utils/analytics';
-import {getDuration} from 'sentry/utils/duration/getDuration';
-import {useOrganization} from 'sentry/utils/useOrganization';
-import {MessageToolCalls} from 'sentry/views/explore/conversations/components/messageToolCalls';
-import type {ConversationMessage} from 'sentry/views/explore/conversations/utils/conversationMessages';
-import {extractMessagesFromNodes} from 'sentry/views/explore/conversations/utils/conversationMessages';
-import {EMPTY_TEXT_CONTENT} from 'sentry/views/insights/pages/agents/utils/aiMessageNormalizer';
-import type {AITraceSpanNode} from 'sentry/views/insights/pages/agents/utils/types';
-import {AIContentRenderer} from 'sentry/views/performance/newTraceDetails/traceDrawer/details/span/eapSections/aiContentRenderer';
-
-interface MessagesPanelProps {
-  nodes: AITraceSpanNode[];
-  onSelectNode: (node: AITraceSpanNode) => void;
-  selectedNodeId: string | null;
-}
-
-export function MessagesPanel({nodes, selectedNodeId, onSelectNode}: MessagesPanelProps) {
-  const organization = useOrganization();
-  const messages = useMemo(() => extractMessagesFromNodes(nodes), [nodes]);
-  const [clickedMessageId, setClickedMessageId] = useState<string | null>(null);
-
-  // Compute effective selected message: use clicked message if it matches current node,
-  // otherwise fall back to assistant message for the selected node
-  const effectiveSelectedMessageId = useMemo(() => {
-    if (clickedMessageId) {
-      const clickedMessage = messages.find(m => m.id === clickedMessageId);
-      if (clickedMessage?.nodeId === selectedNodeId) {
-        return clickedMessageId;
-      }
-    }
-    // Fall back to assistant message for the selected node
-    const assistantMessage = messages.find(
-      m => m.nodeId === selectedNodeId && m.role === 'assistant'
-    );
-    return assistantMessage?.id ?? null;
-  }, [clickedMessageId, messages, selectedNodeId]);
-
-  const nodeMap = useMemo(() => {
-    const map = new Map<string, AITraceSpanNode>();
-    for (const node of nodes) {
-      map.set(node.id, node);
-    }
-    return map;
-  }, [nodes]);
-
-  const handleMessageClick = useCallback(
-    (message: ConversationMessage) => {
-      trackAnalytics('conversations.message.click', {
-        organization,
-      });
-      setClickedMessageId(message.id);
-      const node = nodeMap.get(message.nodeId);
-      if (node) {
-        onSelectNode(node);
-      }
-    },
-    [nodeMap, onSelectNode, organization]
-  );
-
-  if (messages.length === 0) {
-    return (
-      <Stack padding="lg md md md" background="secondary" minHeight="100%" width="100%">
-        <EmptyMessage>{t('No messages found')}</EmptyMessage>
-      </Stack>
-    );
-  }
-
-  return (
-    <Stack padding="lg md md md" background="secondary" minHeight="100%" width="100%">
-      <Stack gap="md" width="100%">
-        {messages.map((message, index) => {
-          const isSelected = message.id === effectiveSelectedMessageId;
-          const isAssistant = message.role === 'assistant';
-          return (
-            <MessageBubble
-              key={index}
-              role={message.role}
-              isClickable={isAssistant}
-              isSelected={isAssistant && isSelected}
-              onClick={isAssistant ? () => handleMessageClick(message) : undefined}
-            >
-              <MessageHeader role={message.role}>
-                {message.role === 'user' ? (
-                  <Text bold size="sm">
-                    {message.userEmail || t('User')}
-                  </Text>
-                ) : (
-                  <Flex align="baseline" gap="sm" flex={1}>
-                    <Text bold size="sm">
-                      {message.agentName || message.modelName || t('Assistant')}
-                    </Text>
-                    {message.duration !== undefined && message.duration > 0 && (
-                      <Text size="xs" variant="muted">
-                        {getDuration(message.duration, 1, true)}
-                      </Text>
-                    )}
-                  </Flex>
-                )}
-              </MessageHeader>
-              {isAssistant && message.toolCalls && message.toolCalls.length > 0 && (
-                <MessageToolCalls
-                  toolCalls={message.toolCalls}
-                  selectedNodeId={selectedNodeId}
-                  nodeMap={nodeMap}
-                  onSelectNode={onSelectNode}
-                />
-              )}
-              {isAssistant && message.reasoning && (
-                <ReasoningSection reasoning={message.reasoning} />
-              )}
-              {message.content === EMPTY_TEXT_CONTENT ? (
-                <Container padding="md">
-                  <MessageText size="sm" align="left" variant="muted">
-                    {message.content}
-                  </MessageText>
-                </Container>
-              ) : (
-                message.content !== '' && (
-                  <StyledClippedBox
-                    clipHeight={200}
-                    buttonProps={{variant: 'secondary', size: 'xs'}}
-                    collapsible
-                  >
-                    <Container padding="md">
-                      <MessageText size="sm" align="left">
-                        <AIContentRenderer
-                          text={message.content}
-                          inline
-                          autoCollapseLimit={10}
-                        />
-                      </MessageText>
-                    </Container>
-                  </StyledClippedBox>
-                )
-              )}
-            </MessageBubble>
-          );
-        })}
-      </Stack>
-    </Stack>
-  );
-}
-
-const MessageHeader = styled('div')<{role: 'user' | 'assistant'}>`
-  display: flex;
-  align-items: center;
-  gap: ${p => p.theme.space.sm};
-  padding: ${p => p.theme.space.sm} ${p => p.theme.space.md};
-  padding-bottom: ${p => (p.role === 'user' ? '0' : p.theme.space.sm)};
-  justify-content: flex-start;
-
-  ${p =>
-    p.role === 'assistant' &&
-    css`
-      position: relative;
-      &::after {
-        content: '';
-        position: absolute;
-        left: ${p.theme.space.md};
-        right: ${p.theme.space.md};
-        bottom: 0;
-        border-bottom: 1px solid ${p.theme.tokens.border.primary};
-      }
-    `}
-`;
-
-const MessageText = styled(Text)`
-  word-break: break-word;
-`;
-
-const MessageBubble = styled('div')<{
-  role: 'user' | 'assistant';
-  isClickable?: boolean;
-  isSelected?: boolean;
-}>`
-  position: relative;
-  z-index: 0;
-  border-radius: ${p => p.theme.radius.md};
-  overflow: hidden;
-  width: 100%;
-  align-self: flex-start;
-
-  ${p =>
-    p.role === 'assistant'
-      ? css`
-          background-color: ${p.theme.tokens.background.primary};
-          &::after {
-            content: '';
-            position: absolute;
-            inset: 0;
-            border: 1px solid ${p.theme.tokens.border.primary};
-            border-radius: inherit;
-            box-sizing: border-box;
-            z-index: 1;
-            pointer-events: none;
-          }
-        `
-      : ''}
-
-  ${p =>
-    p.isClickable &&
-    css`
-      cursor: pointer;
-      &:hover::after {
-        border-color: ${p.theme.tokens.border.accent.moderate};
-        border-width: 2px;
-      }
-    `}
-  ${p =>
-    p.isSelected &&
-    css`
-      &::after {
-        border-color: ${p.theme.tokens.focus.default};
-        border-width: 2px;
-      }
-      &:hover::after {
-        border-color: ${p.theme.tokens.focus.default};
-      }
-    `}
-`;
-
-function ReasoningSection({reasoning}: {reasoning: string}) {
-  const organization = useOrganization();
-  const [isOpen, setIsOpen] = useState(false);
-
-  const handleToggle = (open: boolean) => {
-    setIsOpen(open);
-    trackAnalytics('conversations.detail.expand-thinking', {
-      organization,
-      expanded: open,
-    });
-  };
-
-  // A collapsed <details> keeps its text in the DOM so find-in-page can reveal it.
-  return (
-    <ReasoningDetails onToggle={e => handleToggle(e.currentTarget.open)}>
-      <Flex
-        as="summary"
-        align="center"
-        gap="xs"
-        padding="sm md 0"
-        width="100%"
-        justify="start"
-        cursor="pointer"
-        onClick={e => e.stopPropagation()}
-      >
-        <Text size="xs" variant="muted" monospace italic>
-          {t('Thinking...')}
-        </Text>
-        <IconChevron direction={isOpen ? 'down' : 'right'} size="xs" variant="muted" />
-      </Flex>
-      <Container padding="md">
-        <MessageText size="sm" align="left" variant="muted" monospace italic>
-          <AIContentRenderer text={reasoning} inline autoCollapseLimit={10} />
-        </MessageText>
-      </Container>
-    </ReasoningDetails>
-  );
-}
-
-const ReasoningDetails = styled('details')`
-  width: 100%;
-
-  summary {
-    list-style: none;
-  }
-  summary::-webkit-details-marker {
-    display: none;
-  }
-`;
-
-const StyledClippedBox = styled(ClippedBox)`
-  padding: 0;
-`;
