+++ static/app/views/explore/conversations/hooks/useConversation.tsx
@@ -9,15 +9,12 @@ import {normalizeDateTimeParams} from 'sentry/components/pageFilters/parse';
 import {usePageFilters} from 'sentry/components/pageFilters/usePageFilters';
 import {apiOptions} from 'sentry/utils/api/apiOptions';
 import {useOrganization} from 'sentry/utils/useOrganization';
-import {hasGenAiConversationsRedesignFeature} from 'sentry/views/explore/conversations/utils/features';
 import {getGenAiOperationTypeFromSpanName} from 'sentry/views/insights/pages/agents/utils/query';
 import type {AITraceSpanNode} from 'sentry/views/insights/pages/agents/utils/types';
 import {SpanFields} from 'sentry/views/insights/types';
-import {EAPSpanNodeDetails} from 'sentry/views/performance/newTraceDetails/traceDrawer/details/span';
 import {AiSpanDetails} from 'sentry/views/performance/newTraceDetails/traceDrawer/details/span/aiSpanDetails';
 import type {TraceTreeNodeDetailsProps} from 'sentry/views/performance/newTraceDetails/traceDrawer/tabs/traceTreeNodeDetails';
 import type {TraceTree} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
-import type {EapSpanNode} from 'sentry/views/performance/newTraceDetails/traceModels/traceTreeNode/eapSpanNode';
 
 export interface UseConversationsOptions {
   conversationId: string;
@@ -81,7 +78,7 @@ interface UseConversationResult {
 
 /**
  * Creates a node-like object from a flat API span response so existing UI
- * components (AISpanList, MessagesPanel) work without full trace fetches.
+ * components (AiSpanTimeline, MessagesPanelNew) work without full trace fetches.
  */
 function createNodeFromApiSpan(
   apiSpan: ConversationApiSpan,
@@ -198,10 +195,7 @@ function createNodeFromApiSpan(
     findParentEapTransaction: () => null,
 
     renderDetails(props: TraceTreeNodeDetailsProps<AITraceSpanNode>) {
-      if (hasGenAiConversationsRedesignFeature(props.organization)) {
-        return <AiSpanDetails node={props.node} traceId={props.traceId} />;
-      }
-      return <EAPSpanNodeDetails {...props} node={this as unknown as EapSpanNode} />;
+      return <AiSpanDetails node={props.node} traceId={props.traceId} />;
     },
   };
