+++ static/app/views/explore/conversations/layout.spec.tsx
@@ -7,143 +7,76 @@ import {TopBar} from 'sentry/views/navigation/topBar';
 
 import ConversationsLayout from './layout';
 
+const organization = OrganizationFixture({
+  features: ['performance-view', 'gen-ai-conversations'],
+});
+
+function renderLayout(
+  location: {pathname: string; query?: Record<string, string | number | string[]>},
+  route: string
+) {
+  return render(
+    <TopBar.Slot.Provider>
+      <div data-test-id="top-bar-container">
+        <TopBar />
+      </div>
+      <ConversationsLayout />
+    </TopBar.Slot.Provider>,
+    {
+      organization,
+      initialRouterConfig: {
+        route,
+        location,
+      },
+    }
+  );
+}
+
 describe('ConversationsLayout', () => {
   beforeEach(() => {
     PageFiltersStore.init();
   });
 
-  it('renders detail breadcrumbs in the top bar', async () => {
-    const organization = OrganizationFixture({
-      features: ['performance-view', 'gen-ai-conversations'],
-    });
-
-    render(
-      <TopBar.Slot.Provider>
-        <div data-test-id="top-bar-container">
-          <TopBar />
-        </div>
-        <ConversationsLayout />
-      </TopBar.Slot.Provider>,
-      {
-        organization,
-        initialRouterConfig: {
-          route: '/organizations/:orgId/explore/conversations/:conversationId/',
-          location: {
-            pathname: `/organizations/${organization.slug}/explore/conversations/6c5b72fc/`,
-            query: {
-              environment: ['prod'],
-              project: ['1'],
-              statsPeriod: '7d',
-            },
-          },
-        },
-      }
+  it('renders the landing title on the list page', async () => {
+    renderLayout(
+      {pathname: `/organizations/${organization.slug}/explore/conversations/`},
+      '/organizations/:orgId/explore/conversations/'
     );
 
     const topBar = screen.getByTestId('top-bar-container');
-    expect(await within(topBar).findByText('6c5b72fc')).toBeInTheDocument();
-    expect(within(topBar).getByRole('link', {name: 'Conversations'})).toHaveAttribute(
-      'href',
-      `/organizations/${organization.slug}/explore/conversations/?environment=prod&project=1&referrer=conversations-breadcrumb&statsPeriod=24h`
-    );
+    expect(await within(topBar).findByText('AI Conversations')).toBeInTheDocument();
   });
 
-  it('restores the originating list filters from router state', async () => {
-    const organization = OrganizationFixture({
-      features: ['performance-view', 'gen-ai-conversations'],
-    });
-
-    render(
-      <TopBar.Slot.Provider>
-        <div data-test-id="top-bar-container">
-          <TopBar />
-        </div>
-        <ConversationsLayout />
-      </TopBar.Slot.Provider>,
+  it('renders saved query breadcrumbs on the list page', async () => {
+    renderLayout(
       {
-        organization,
-        initialRouterConfig: {
-          route: '/organizations/:orgId/explore/conversations/:conversationId/',
-          location: {
-            pathname: `/organizations/${organization.slug}/explore/conversations/6c5b72fc/`,
-            // The detail page repurposes start/end as the conversation window,
-            // so the list filters are carried over via router state instead.
-            query: {
-              start: '2024-01-01T00:00:00.000Z',
-              end: '2024-01-02T00:00:00.000Z',
-              project: ['1'],
-            },
-            state: {
-              conversationsListQuery: {
-                project: ['1'],
-                environment: ['prod'],
-                statsPeriod: '7d',
-                query: 'gen_ai.conversation.id:abc',
-                agent: 'my-agent',
-              },
-            },
-          },
-        },
-      }
-    );
-
-    const topBar = screen.getByTestId('top-bar-container');
-    const link = await within(topBar).findByRole('link', {name: 'Conversations'});
-    const href = link.getAttribute('href')!;
-    const [pathname, search] = href.split('?');
-    const params = new URLSearchParams(search);
-
-    expect(pathname).toBe(`/organizations/${organization.slug}/explore/conversations/`);
-    // The user's original list filters are restored exactly...
-    expect(params.get('statsPeriod')).toBe('7d');
-    expect(params.get('environment')).toBe('prod');
-    expect(params.get('project')).toBe('1');
-    expect(params.get('query')).toBe('gen_ai.conversation.id:abc');
-    expect(params.get('agent')).toBe('my-agent');
-    // ...and the conversation-scoped time window does not leak back in.
-    expect(params.has('start')).toBe(false);
-    expect(params.has('end')).toBe(false);
-  });
-
-  it('refreshes restored filters when re-opening the same conversation', async () => {
-    const organization = OrganizationFixture({
-      features: ['performance-view', 'gen-ai-conversations'],
-    });
-    const detailPathname = `/organizations/${organization.slug}/explore/conversations/6c5b72fc/`;
-
-    const {router} = render(
-      <TopBar.Slot.Provider>
-        <div data-test-id="top-bar-container">
-          <TopBar />
-        </div>
-        <ConversationsLayout />
-      </TopBar.Slot.Provider>,
-      {
-        organization,
-        initialRouterConfig: {
-          route: '/organizations/:orgId/explore/conversations/:conversationId/',
-          location: {
-            pathname: detailPathname,
-            state: {conversationsListQuery: {statsPeriod: '7d'}},
-          },
-        },
-      }
+        pathname: `/organizations/${organization.slug}/explore/conversations/`,
+        query: {id: 'abc', title: 'My saved query'},
+      },
+      '/organizations/:orgId/explore/conversations/'
     );
 
     const topBar = screen.getByTestId('top-bar-container');
     expect(
       await within(topBar).findByRole('link', {name: 'Conversations'})
-    ).toHaveAttribute('href', expect.stringContaining('statsPeriod=7d'));
+    ).toBeInTheDocument();
+    expect(within(topBar).getByText('My saved query')).toBeInTheDocument();
+  });
 
-    // Re-open the same conversation with a different list query (e.g. the user
-    // changed filters on the list before clicking back in).
-    router.navigate(detailPathname, {
-      state: {conversationsListQuery: {statsPeriod: '14d', query: 'foo:bar'}},
-    });
+  it('defers the title to the detail page on a conversation detail route', async () => {
+    renderLayout(
+      {
+        pathname: `/organizations/${organization.slug}/explore/conversations/6c5b72fc/`,
+      },
+      '/organizations/:orgId/explore/conversations/:conversationId/'
+    );
 
-    const link = await within(topBar).findByRole('link', {name: 'Conversations'});
-    const params = new URLSearchParams(link.getAttribute('href')!.split('?')[1]);
-    expect(params.get('statsPeriod')).toBe('14d');
-    expect(params.get('query')).toBe('foo:bar');
+    // The detail page renders its own breadcrumbs, so the layout leaves the
+    // top bar title slot empty.
+    const topBar = await screen.findByTestId('top-bar-container');
+    expect(within(topBar).queryByText('AI Conversations')).not.toBeInTheDocument();
+    expect(
+      within(topBar).queryByRole('link', {name: 'Conversations'})
+    ).not.toBeInTheDocument();
   });
 });
