+++ static/app/views/explore/conversations/layout.tsx
@@ -1,10 +1,8 @@
-import {Fragment, useRef} from 'react';
+import {Fragment} from 'react';
 import {Outlet} from 'react-router-dom';
-import type {Location} from 'history';
 
 import {FeatureBadge} from '@sentry/scraps/badge';
 import {Stack} from '@sentry/scraps/layout';
-import {Tooltip} from '@sentry/scraps/tooltip';
 
 import Feature from 'sentry/components/acl/feature';
 import {AnalyticsArea} from 'sentry/components/analyticsArea';
@@ -16,7 +14,6 @@ import {PageFiltersContainer} from 'sentry/components/pageFilters/container';
 import {SentryDocumentTitle} from 'sentry/components/sentryDocumentTitle';
 import {defined} from 'sentry/utils/defined';
 import {decodeScalar} from 'sentry/utils/queryString';
-import {isUUID} from 'sentry/utils/string/isUUID';
 import {normalizeUrl} from 'sentry/utils/url/normalizeUrl';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useOrganization} from 'sentry/utils/useOrganization';
@@ -27,8 +24,6 @@ import {
   CONVERSATIONS_SIDEBAR_LABEL,
   MAX_PICKABLE_DAYS,
 } from 'sentry/views/explore/conversations/settings';
-import {hasGenAiConversationsRedesignFeature} from 'sentry/views/explore/conversations/utils/features';
-import {getConversationsListQueryFromState} from 'sentry/views/explore/conversations/utils/listNavigation';
 import {TopBar} from 'sentry/views/navigation/topBar';
 
 function ConversationsLayout() {
@@ -76,100 +71,22 @@ function ConversationsLayoutContent() {
 }
 
 function ConversationsHeader() {
-  const organization = useOrganization();
-  const location = useLocation();
   const {conversationId} = useParams<{conversationId?: string}>();
-
   const isDetailPage = !!conversationId;
-  // The redesigned detail page renders its own breadcrumbs (with the project
-  // badge and copy affordance) into the title slot, so the layout only owns
-  // the breadcrumbs for the landing and legacy detail pages.
-  const isRedesign = hasGenAiConversationsRedesignFeature(organization);
-  const conversationsBaseUrl = normalizeUrl(
-    `/organizations/${organization.slug}/explore/${CONVERSATIONS_LANDING_SUB_PATH}/`
-  );
 
-  // The list location we navigated from is passed via router state so the
-  // breadcrumb can return to the exact filtered list (mirroring browser
-  // "back"). Cache it per-conversation so a later in-page navigation that
-  // drops the state (e.g. selecting a span) doesn't lose it.
-  const restoredListQuery = useRestoredListQuery(conversationId, location.state);
-
-  const backToListCrumb = restoredListQuery
-    ? {
-        pathname: conversationsBaseUrl,
-        query: {...restoredListQuery, referrer: 'conversations-breadcrumb'},
-      }
-    : {
-        pathname: conversationsBaseUrl,
-        query: {
-          statsPeriod: '24h',
-          start: undefined,
-          end: undefined,
-          referrer: 'conversations-breadcrumb',
-        },
-      };
-
+  // The detail page renders its own breadcrumbs (with the project badge and
+  // copy affordance) into the title slot, so the layout only owns the title
+  // for the landing page.
   return (
     <Fragment>
       <TopBar.Slot name="title">
-        {isDetailPage && isRedesign ? null : isDetailPage ? (
-          <Breadcrumbs
-            crumbs={[
-              {
-                label: CONVERSATIONS_SIDEBAR_LABEL,
-                to: backToListCrumb,
-                // When we have the originating list query it already holds the
-                // full filter state; preserving page filters would merge the
-                // detail page's conversation-scoped start/end on top of it.
-                preservePageFilters: !restoredListQuery,
-              },
-              {
-                label: isUUID(conversationId) ? (
-                  conversationId.slice(0, 8)
-                ) : (
-                  <Tooltip title={conversationId}>
-                    <span>{conversationId}</span>
-                  </Tooltip>
-                ),
-              },
-            ]}
-          />
-        ) : (
-          <ConversationsLandingTitle />
-        )}
+        {isDetailPage ? null : <ConversationsLandingTitle />}
       </TopBar.Slot>
       <TopBar.Slot name="feedback">
         <FeedbackButton>{null}</FeedbackButton>
       </TopBar.Slot>
     </Fragment>
   );
-}
-
-/**
- * Returns the originating list querystring for the current conversation, read
- * from router location state. Caches it per-conversation so an in-page
- * navigation that clears the state (e.g. nuqs `replace` when selecting a span)
- * keeps the breadcrumb pointing back at the filtered list.
- */
-function useRestoredListQuery(
-  conversationId: string | undefined,
-  state: Location['state']
-): Location['query'] | undefined {
-  const cache = useRef<{conversationId?: string; query?: Location['query']}>({});
-  const listQueryFromState = getConversationsListQueryFromState(state);
-
-  // Fresh state always wins, so re-opening the same conversation with new
-  // filters refreshes the cache. When the state is absent (an in-page nuqs
-  // `replace`), keep the cached query as long as we're on the same
-  // conversation; otherwise fall back to the default.
-  if (listQueryFromState) {
-    cache.current = {conversationId, query: listQueryFromState};
-  }
-
-  return cache.current.conversationId === conversationId
-    ? cache.current.query
-    : undefined;
 }
 
 function ConversationsLandingTitle() {
