+++ static/app/views/explore/conversations/overview.tsx
@@ -23,13 +23,11 @@ import {
   ExploreBodySearch,
 } from 'sentry/views/explore/components/styles';
 import {TraceItemSearchQueryBuilder} from 'sentry/views/explore/components/traceItemSearchQueryBuilder';
-import {ConversationsTable} from 'sentry/views/explore/conversations/components/conversationsTable';
 import {ConversationsTableNew} from 'sentry/views/explore/conversations/components/conversationsTableNew';
 import {SaveConversationQueryButton} from 'sentry/views/explore/conversations/components/saveConversationQueryButton';
 import {useShowConversationOnboarding} from 'sentry/views/explore/conversations/hooks/useShowConversationOnboarding';
 import {ConversationOnboarding} from 'sentry/views/explore/conversations/onboarding';
 import {MAX_PICKABLE_DAYS} from 'sentry/views/explore/conversations/settings';
-import {hasGenAiConversationsRedesignFeature} from 'sentry/views/explore/conversations/utils/features';
 import {AgentSelector} from 'sentry/views/insights/common/components/agentSelector';
 import {useTableCursor} from 'sentry/views/insights/pages/agents/hooks/useTableCursor';
 import {
@@ -97,8 +95,6 @@ function ConversationsOverviewPage() {
   const {spanSearchQueryBuilderProviderProps, spanSearchQueryBuilderProps} =
     useSpanSearchQueryBuilderProps(searchQueryBuilderProps);
 
-  const showConversationsTableNew = hasGenAiConversationsRedesignFeature(organization);
-
   return (
     <SearchQueryBuilderProvider {...spanSearchQueryBuilderProviderProps}>
       <ExploreBodySearch>
@@ -132,20 +128,13 @@ function ConversationsOverviewPage() {
         </Layout.Main>
       </ExploreBodySearch>
       <ExploreBodyContent>
-        <Stack
-          flex={1}
-          minWidth={showConversationsTableNew ? '0' : undefined}
-          padding="xl"
-          gap="md"
-        >
+        <Stack flex={1} minWidth="0" padding="xl" gap="md">
           {isOnboardingLoading ? (
             <LoadingIndicator />
           ) : showOnboarding ? (
             <ConversationOnboarding onDismiss={refetchOnboarding} />
-          ) : showConversationsTableNew ? (
-            <ConversationsTableNew />
           ) : (
-            <ConversationsTable />
+            <ConversationsTableNew />
           )}
         </Stack>
       </ExploreBodyContent>
