+++ static/app/views/performance/newTraceDetails/traceDrawer/tabs/traceAiSpans.tsx
@@ -16,8 +16,6 @@ import {
   ConversationLeftPanel,
   ConversationSplitLayout,
 } from 'sentry/views/explore/conversations/components/conversationLayout';
-import {hasGenAiConversationsRedesignFeature} from 'sentry/views/explore/conversations/utils/features';
-import {AISpanList} from 'sentry/views/insights/pages/agents/components/aiSpanList';
 import {AiSpanTimeline} from 'sentry/views/insights/pages/agents/components/aiSpanTimeline';
 import {useAITrace} from 'sentry/views/insights/pages/agents/hooks/useAITrace';
 import {getDefaultSelectedNode} from 'sentry/views/insights/pages/agents/utils/getDefaultSelectedNode';
@@ -100,8 +98,6 @@ export function TraceAiSpans({
 
   const {selectedNode, handleSelectNode} = useAiSpanSelection(nodes);
 
-  const showTimeline = hasGenAiConversationsRedesignFeature(organization);
-
   const handleViewFullTraceClick = () => {
     trackAnalytics('agent-monitoring.trace.view-full-trace-click', {
       organization,
@@ -140,19 +136,11 @@ export function TraceAiSpans({
       </HeaderCell>
       <LeftPanel>
         <SpansHeader>{t('AI Spans')}</SpansHeader>
-        {showTimeline ? (
-          <AiSpanTimeline
-            nodes={nodes}
-            onSelectNode={handleSelectNode}
-            selectedNodeKey={selectedNode?.id ?? null}
-          />
-        ) : (
-          <AISpanList
-            nodes={nodes}
-            onSelectNode={handleSelectNode}
-            selectedNodeKey={selectedNode?.id ?? null}
-          />
-        )}
+        <AiSpanTimeline
+          nodes={nodes}
+          onSelectNode={handleSelectNode}
+          selectedNodeKey={selectedNode?.id ?? null}
+        />
       </LeftPanel>
       <RightPanel>
         {selectedNode?.renderDetails({
@@ -182,10 +170,7 @@ export function AiSpansSplitView({
   nodes: AITraceSpanNode[];
   traceSlug: string;
 }) {
-  const organization = useOrganization();
   const {selectedNode, handleSelectNode} = useAiSpanSelection(nodes);
-
-  const showTimeline = hasGenAiConversationsRedesignFeature(organization);
 
   const nodeTraceMap = useMemo(() => {
     const map = new Map<string, string>();
@@ -205,19 +190,11 @@ export function AiSpansSplitView({
         <ConversationLeftPanel>
           <Flex flex="1" minHeight="0" overflowY="auto" overflowX="hidden">
             <Container padding="md lg md lg" width="100%">
-              {showTimeline ? (
-                <AiSpanTimeline
-                  nodes={nodes}
-                  onSelectNode={handleSelectNode}
-                  selectedNodeKey={selectedNode?.id ?? null}
-                />
-              ) : (
-                <AISpanList
-                  nodes={nodes}
-                  onSelectNode={handleSelectNode}
-                  selectedNodeKey={selectedNode?.id ?? null}
-                />
-              )}
+              <AiSpanTimeline
+                nodes={nodes}
+                onSelectNode={handleSelectNode}
+                selectedNodeKey={selectedNode?.id ?? null}
+              />
             </Container>
           </Flex>
         </ConversationLeftPanel>
